requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pgvector
sentence-transformers
langchain-groq
langchain-mistralai
langgraph
requests
fpdf
reportlab
python-multipart